+++ app/src/main/java/com/example/testwithpoetry/domain/models/Author.kt
@@ -1,4 +1,4 @@
-package com.example.testwithpoetry.localModels
+package com.example.testwithpoetry.domain.models
 
 data class Author(
     val name: String,
