+++ app/src/main/java/com/example/testwithpoetry/ui/screens/AuthorDetailScreen.kt
@@ -1,4 +1,4 @@
-package com.example.testwithpoetry.ui.screen
+package com.example.testwithpoetry.ui.screens
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -22,7 +22,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.example.testwithpoetry.AuthorDetailViewModel
+import com.example.testwithpoetry.ui.viewmodel.AuthorDetailViewModel
 import com.example.testwithpoetry.ui.theme.TestWithPoetryTheme
 
 @Composable
